Issue #4686: Moved input of 'flawless' subdirectory to 'checkstyleanttask' subdirectory of CheckstyleAntTaskTest
